Remove deprecated BuildCacheKey.getDisplayName()
And also remove DefaultBuildCacheKey and replace it with SimpleBuildCacheKey.
