+++ platforms/core-execution/build-cache/src/test/groovy/org/gradle/caching/internal/controller/DefaultBuildCacheControllerPackOperationExecutorTest.groovy
@@ -19,6 +19,7 @@ package org.gradle.caching.internal.controller
 import groovy.transform.Immutable
 import org.gradle.api.internal.cache.StringInterner
 import org.gradle.caching.internal.CacheableEntity
+import org.gradle.caching.internal.SimpleBuildCacheKey
 import org.gradle.caching.internal.controller.operations.PackOperationResult
 import org.gradle.caching.internal.controller.operations.UnpackOperationResult
 import org.gradle.caching.internal.origin.OriginMetadata
@@ -26,7 +27,6 @@ import org.gradle.caching.internal.origin.OriginMetadataFactory
 import org.gradle.caching.internal.origin.OriginReader
 import org.gradle.caching.internal.origin.OriginWriter
 import org.gradle.caching.internal.packaging.BuildCacheEntryPacker
-import org.gradle.internal.execution.caching.impl.DefaultBuildCacheKey
 import org.gradle.internal.file.FileMetadata
 import org.gradle.internal.file.TreeType
 import org.gradle.internal.file.impl.DefaultFileMetadata
@@ -60,7 +60,7 @@ class DefaultBuildCacheControllerPackOperationExecutorTest extends Specification
     def buildOperationRunner = Mock(BuildOperationRunner)
     PackOperationExecutor packOperationExecutor = new PackOperationExecutor(buildOperationRunner, packer, originFactory, stringInterner)
 
-    def key = new DefaultBuildCacheKey(TestHashCodes.hashCodeFrom(1234))
+    def key = new SimpleBuildCacheKey(TestHashCodes.hashCodeFrom(1234))
 
     def originMetadata = Mock(OriginMetadata)
     def originReader = Mock(OriginReader)
