+++ testing/architecture-test/src/changes/accepted-changes/accepted-public-api-changes.json
@@ -1017,6 +1017,14 @@
             ]
         },
         {
+            "type": "org.gradle.caching.BuildCacheKey",
+            "member": "Method org.gradle.caching.BuildCacheKey.getDisplayName()",
+            "acceptation": "Removed deprecated method",
+            "changes": [
+                "Method has been removed"
+            ]
+        },
+        {
             "type": "org.gradle.jvm.application.tasks.CreateStartScripts",
             "member": "Method org.gradle.jvm.application.tasks.CreateStartScripts.getMainClassName()",
             "acceptation": "[ADD YOUR CUSTOM REASON HERE]",
